Fixed quotes issue and expand, norminette is good too
