+++ src/lexer/expander.c
@@ -6,7 +6,7 @@
 /*   By: rraumain <rraumain@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/02/23 13:43:49 by rraumain          #+#    #+#             */
-/*   Updated: 2025/02/25 15:11:10 by rraumain         ###   ########.fr       */
+/*   Updated: 2025/02/26 18:26:07 by rraumain         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -51,7 +51,7 @@ static char	*get_var_value(char *varname, t_global_data *data)
 	return (value);
 }
 
-char	*expand_var(char *token, int *index, t_global_data *data)
+static char	*expand_var(char *token, int *index, t_global_data *data)
 {
 	char	*varname;
 	char	*value;
@@ -71,4 +71,23 @@ char	*expand_var(char *token, int *index, t_global_data *data)
 		return (token);
 	*index += ft_strlen(value);
 	return (expanded);
+}
+
+void	expand_word(char **word, t_global_data *data)
+{
+	char	*tmp;
+	int		i;
+
+	i = 0;
+	while ((*word)[i])
+	{
+		if ((*word)[i] == '$')
+		{
+			tmp = expand_var(*word, &i, data);
+			free(*word);
+			*word = tmp;
+		}
+		else
+			i++;
+	}
 }
